refactor: barrel test file with single skipIf guard for docker tests
Move test logic into plain .ts modules (entrypoint.ts, tools.ts) that
export registration functions. A single docker.test.ts barrel imports
them and wraps everything in one describe.skipIf(!IN_CONTAINER) block.
